feat(keystore_and_bump): Fixed things
