+++ pragma-sdk/pragma_sdk/onchain/types/types.py
@@ -3,8 +3,7 @@ from dataclasses import dataclass
 from enum import StrEnum, unique
 from collections import namedtuple
 
-from pydantic import HttpUrl, model_validator
-from starknet_py.net.client_models import ResourceBounds
+from pydantic import HttpUrl
 from starknet_py.net.client import Tag as BlockTag
 from starknet_py.contract import InvokeResult
 
@@ -38,20 +37,6 @@ PublishEntriesOnChainResult = List[InvokeResult]
 
 BlockNumber = int
 BlockId = Union[BlockTag, BlockNumber]
-
-
-@dataclass(frozen=True)
-class ExecutionConfig:
-    pagination: int = 40
-    max_fee: int = int(1e18)
-    enable_strk_fees: bool = False
-    l1_resource_bounds: Optional[ResourceBounds] = None
-    auto_estimate: bool = False
-
-    @model_validator(mode="after")  # type: ignore[misc]
-    def post_root(self) -> None:
-        if self.auto_estimate == (self.l1_resource_bounds is not None):
-            raise ValueError("Either auto_estimate or l1_resource_bounds must be set")
 
 
 @unique
