+++ price-pusher/price_pusher/main.py
@@ -2,7 +2,7 @@ import asyncio
 import click
 import logging
 
-from typing import Optional, List
+from typing import Optional, List, Tuple
 
 from pragma_sdk.common.fetchers.fetcher_client import FetcherClient
 from pragma_sdk.common.types.client import PragmaClient
@@ -32,7 +32,7 @@ async def main(
     price_configs: List[PriceConfig],
     target: Target,
     network: Network,
-    private_key: str,
+    private_key: str | Tuple[str, str],
     publisher_name: str,
     publisher_address: str,
     rpc_url: Optional[str] = None,
@@ -104,7 +104,7 @@ def _create_client(
     target: Target,
     network: Network,
     publisher_address: str,
-    private_key: str,
+    private_key: str | Tuple[str, str],
     rpc_url: Optional[str] = None,
     api_base_url: Optional[str] = None,
     api_key: Optional[str] = None,
@@ -194,6 +194,7 @@ def _create_client(
 @click.option(
     "-p",
     "--private-key",
+    "raw_private_key",
     type=click.STRING,
     required=True,
     help=(
@@ -247,7 +248,7 @@ def cli_entrypoint(
     target: str,
     network: str,
     rpc_url: Optional[str],
-    private_key: str,
+    raw_private_key: str,
     publisher_name: str,
     publisher_address: str,
     api_base_url: Optional[str],
@@ -287,7 +288,7 @@ def cli_entrypoint(
     sdk_logger.setLevel(log_level)
 
     setup_logging(logger, log_level)
-    private_key = load_private_key_from_cli_arg(private_key)
+    private_key = load_private_key_from_cli_arg(raw_private_key)
     price_configs: List[PriceConfig] = PriceConfig.from_yaml(config_file)
 
     # Make sure that the API base url does not ends with /
