+++ vrf-listener/vrf_listener/main.py
@@ -28,8 +28,8 @@ async def main(
     client = PragmaOnChainClient(
         chain_name=network,
         network=network if rpc_url is None else rpc_url,
-        account_contract_address=int(admin_address, 16),
-        account_private_key=int(private_key, 16),
+        account_contract_address=admin_address,
+        account_private_key=private_key,
         contract_addresses_config=ContractAddresses(
             publisher_registry_address=0x0,
             oracle_proxy_addresss=int(oracle_address, 16),
@@ -89,14 +89,14 @@ async def main(
 @click.option(
     "-p",
     "--private-key",
+    "raw_private_key",
     type=click.STRING,
     required=True,
     help=(
         "Private key of the signer. Format: "
         "aws:secret_name, "
         "plain:private_key, "
-        "env:ENV_VAR_NAME, "
-        "or keystore:PATH/TO/THE/KEYSTORE:PASSWORD"
+        "or env:ENV_VAR_NAME"
     ),
 )
 @click.option(
@@ -122,7 +122,7 @@ def cli_entrypoint(
     oracle_address: str,
     vrf_address: str,
     admin_address: str,
-    private_key: str,
+    raw_private_key: str,
     start_block: int,
     check_requests_interval: int,
 ) -> None:
@@ -130,7 +130,9 @@ def cli_entrypoint(
     VRF Listener entry point.
     """
     setup_logging(logger, log_level)
-    private_key = load_private_key_from_cli_arg(private_key)
+    private_key = load_private_key_from_cli_arg(raw_private_key)
+    if isinstance(private_key, tuple):
+        raise ValueError("⛔ KeyStores aren't supported as private key for the vrf-listener!")
     asyncio.run(
         main(
             network=network,
